requirements.txt
python-telegram-bot==20.6
aiohttp
python-dotenv==1.0.1
natasha